% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{runApp}
\alias{runApp}
\title{runApp}
\usage{
runApp()
}
\description{
Walks user through outliers processing on a given, preprocessed data package and generates report.

This is a Shiny app that takes input data from the global environment (either matrix or dataframe),
and gives a step by step approach to outlier processing specifically:
1. Ordination Analysis
2. (Generalized) Minimum Covariance Determinant (G)MCD Analysis
3. CorrMax
This app is able to generate a report with the information up until the current step for each step

The server logic is organized as follows:
- Reactive Vals and Reactives: the step#done() reactiveVals are necessary for bookmark restoration
- Observers
- Render Outputs
- App State Trigger Funcs
}
\author{
Logan Lim Jeremy Tanuan
}
